fix: update api path
